refactor(store): split blazegraph-harden into harden/ modules behind a thin facade
The harden migration had grown past 1000 lines in one file, coupling
docker-inspect parsing, plan rendering, executor phases, SPARQL
verification probes, and the rollback routine — an incident fix to any
one of them meant re-reading all of them. Split by responsibility:

  harden/state.ts    — inspect-derived state machine (+ parseInspect,
                       HARDEN_BACKUP_SUFFIX: backup naming is a state
                       concept, and keeping it here keeps the module
                       graph acyclic: steps -> state, never back)
  harden/steps.ts    — step defs + plan (single source of every argv)
  harden/executor.ts — execution/resume, predicate checks, harden lock
  harden/verify.ts   — ASK + identity-tag probes
  harden/rollback.ts — stop-at-first-failure rollback

blazegraph-harden.ts stays as the public facade re-exporting the same
surface (executeHardenMigration, planHardenMigration, hardenStepDefs,
inspectHardenState, constants, types) and carries the incident
narrative. Pure move — no behavior change intended or made; the entire
existing test suite still imports ONLY through the facade and passes
unchanged, which doubles as the proof the facade is complete.

Review thread: blazegraph-harden.ts:507 (decompose the harden module
before it lands over 1k lines).
